% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVAdot.R
\name{ANEVAdot}
\alias{ANEVAdot}
\title{ANEVA Dosage Outlier Test}
\usage{
ANEVAdot(filepath, output_columns = c("refCount", "altCount"),
  eh1 = "refCount", eh2 = "altCount", Eg_std, FDR = 0.05, plot = TRUE)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE count data filepath.
Data should be a raw, tab-delimited text file.}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{Eg_std}{a vector containing standard deviation in the log2 transformed total gene
expression in a healthy population. Eg_std vector must be in one-to-one correspondence with
ASE count data, and must be ordered correctly.}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}

\item{plot}{A logical T/F indicating whether plots should be generated for the test data}

\item{output_columnns}{A vector containing strings of the exact column names to be duplicated
from input file in the output}
}
\value{
A table containing the user-specified output columns, as well as unadjusted and
Benjamini-Hoschberg adjusted p-values for detection of potential dosage outlier. P-values
will not be generated for records with missing or infinite variances, or for those for
which the total allelic count is less than 8.
}
\description{
This test is designed to detect if ASE data reveals sufficient imbalance to induce
an outlier in total gene expression.
}
\examples{
result<-ANEVAdot("data/testdata.txt", output_columns = c("eh1","eh2"), eh1 = "eh1", eh2 = "eh2", Eg_std=Sg)
}
