% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEbinomial.R
\name{ASEbinomial}
\alias{ASEbinomial}
\title{Perform binomial test on ASE count data}
\usage{
ASEbinomial(filepath, output_columns = c("refCount", "altCount"),
  eh1 = "refCount", eh2 = "altCount", prob = 0.5, FDR = 0.05,
  plot = TRUE)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE data filepath}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{prob}{A numeric value between 0 and 1 which will be passed into the binomial test}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}

\item{plot}{A logical T/F indicating whether plots should be generated for the test data}

\item{output_columnns}{A vector containing strings of the exact column names to be duplicated
from input file in the output}
}
\value{
A table containing the user-specified output columns, as well as unadjusted and
Benjamini-Hoschberg adjusted p-values corresponding to the binomial test.
}
\description{
Perform binomial test on ASE count data
}
\examples{
(tbd)
}
