% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEoutlier.R
\name{dp.UndExp}
\alias{dp.UndExp}
\title{We first implement a helper function to generate the desired integrand
for the Loss of Expression test, assuming that the haplotype with the
higher count is the wild type.}
\usage{
dp.UndExp(r, eh1, eh2, Vg)
}
\arguments{
\item{r}{The ?true reference allele ratio?}

\item{eh1}{An integer value for one of the haplotype counts}

\item{eh2}{An integer value for the second haplotype count}

\item{Vg}{A numeric value for the estimated dosage variance in normal population
(can be gotten from GTEx)}
}
\value{
The integrand for which we wish to integrate over to get
desired p-values
}
\description{
We first implement a helper function to generate the desired integrand
for the Loss of Expression test, assuming that the haplotype with the
higher count is the wild type.
}
\examples{
(tbd)
}
