% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r0.R
\name{get_r0}
\alias{get_r0}
\title{Reference bias due to alignment}
\usage{
get_r0(ASEdat, eh1 = "refCount", eh2 = "altCount")
}
\arguments{
\item{ASEdat}{Dataframe containing reference and alternative count data.}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}
}
\value{
A numeric estimate of the reference bias r0 for the given library.
}
\description{
This function is designed to estimate the reference ratio in the absence of any regulatory
difference; that is, the reference bias due to alignment. This is done by obtaining the
median ratio between eh1 and eh2 across the top 20% expressed loci in the library.
}
\examples{}
