% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVA_DOT.R
\name{integrand}
\alias{integrand}
\title{Function generates the required integrand for the test.}
\usage{
integrand(dE, eh1, eh2, Eg_std, r0, p0, log_BinCoeff)
}
\arguments{
\item{dE}{Variable representing the true change in gene dosage
(variable of integration)}

\item{eh1}{Integer value for count of expression of haplotype1}

\item{eh2}{Integer value for count of expression of haplotype2}

\item{Eg_std}{Vector containing genetic standard deviation, the square root of genetic
variation in regulation in general population in natural log scale (numeric value).}

\item{r0}{The ratio of the eh1 allele (i.e., eh1/(eh1+eh2)) in the absence of any regulatory
difference (reference bias due to alignment). The simplest way to get such an estimate would
be to get the median ratio between eh1 and eh2 across the entire library.}

\item{p0}{An average noise rate p(R->A) or p(A->R), i.e., the probability of seeing an allele
due to noise when it is essentially not there. (For v7: LAMP = 0.0003).}

\item{log_BinCoeff}{a vector of natural log (ln) transformed binomial coefficients 0:N=eh1+eh2.}
}
\value{
The integrand over which we wish to integrate to get
desired p-values
}
\description{
Function generates the required integrand for the test.
}
