% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_BinCoeffs.R
\name{log_BinCoeffs}
\alias{log_BinCoeffs}
\title{This function calculates Binomial coefficients for choosing 0:n out of n,
in log scale. It is useful for precalculating the coefficients and
providing them to "Binom_test_fast.R" or "pdf_binom_fast.R" when they
are redone on the same N, over and over again. This function was
designed by Pejman Mohammadi, 2018, Scripps Research, San Diego, CA.}
\usage{
log_BinCoeffs(n)
}
\arguments{
\item{n}{The desired n for which coefficients will be calculated}
}
\value{
log_z Vector of log scale Binomial coefficients
}
\description{
This function calculates Binomial coefficients for choosing 0:n out of n,
in log scale. It is useful for precalculating the coefficients and
providing them to "Binom_test_fast.R" or "pdf_binom_fast.R" when they
are redone on the same N, over and over again. This function was
designed by Pejman Mohammadi, 2018, Scripps Research, San Diego, CA.
}
