% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_Binom_fast.R
\name{pdf_Binom_fast}
\alias{pdf_Binom_fast}
\title{This function generates the binomial pdf, with the option to provide
Binomial coefficients to it. It is useful when it is recalculated on
the same N, over and over again.
log_BinCoeffs can be precalculated using the function log_BinCoeffs(n)
NEED TO INSERT WARNING IF p OUT OF DOMAIN [0,1]}
\usage{
pdf_Binom_fast(n, p, log_BinCoeffs)
}
\arguments{
\item{n}{The desired n for which the pdf will be calculated}

\item{p}{The Binomial parameter probability of success}

\item{log_BinCoeffs}{Optional vector of log-scale binomial coefficients}
}
\value{
Bin_p The pdf for 0:n, out of n
}
\description{
This function generates the binomial pdf, with the option to provide
Binomial coefficients to it. It is useful when it is recalculated on
the same N, over and over again.
log_BinCoeffs can be precalculated using the function log_BinCoeffs(n)
NEED TO INSERT WARNING IF p OUT OF DOMAIN [0,1]
}
