% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVAdot.R
\name{plot.ANEVAdot}
\alias{plot.ANEVAdot}
\title{Generate plot of ASE data showing significant outliers.}
\usage{
\method{plot}{ANEVAdot}(filepath, eh1 = "refCount", eh2 = "altCount",
  Eg_std, FDR = 0.05)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE data filepath}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{Eg_std}{a vector with variance estimates for dosage distributions associated with count data.
Eg_std must be in one-to-one correspondence with ASE data, and ordered correctly.}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}
}
\description{
Generate plot of ASE data showing significant outliers.
}
\examples{
(tbd)
}
