% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEbinomial.R
\name{plot.ASEbinomial}
\alias{plot.ASEbinomial}
\title{Generate plot of ASE data showing binomial outliers.}
\usage{
\method{plot}{ASEbinomial}(filepath, eh1 = "refCount", eh2 = "altCount",
  prob = 0.5, FDR = 0.05)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE data filepath}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{prob}{A numeric value between 0 and 1 which will be passed into the binomial test}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}
}
\description{
Generate plot of ASE data showing binomial outliers.
}
\examples{
(tbd)
}
