% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEoutlier.R
\name{plot.ASEoutlier}
\alias{plot.ASEoutlier}
\title{Generate plot of ASE data showing significant outliers.}
\usage{
\method{plot}{ASEoutlier}(filepath, r, eh1 = "refCount", eh2 = "altCount",
  Vg, FDR = 0.05)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE data filepath}

\item{r}{The ?true reference allele ratio?}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{Vg}{A numeric value for the estimated dosage variance in normal population}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}
}
\description{
Generate plot of ASE data showing significant outliers.
}
\examples{
(tbd)
}
