% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVA_DOT.R
\name{Test_ASE_Outliers}
\alias{Test_ASE_Outliers}
\title{Test ASE Outliers}
\usage{
Test_ASE_Outliers(Eg_std, eh1, eh2, r0, p0)
}
\arguments{
\item{Eg_std}{Standard deviation in the total gene expression in a healthy population.}

\item{eh1}{Integer value for of expression of haplotype 1.}

\item{eh2}{Integer value of expression of haplotype 2.}

\item{r0}{The ratio of the eh1 allele (i.e., eh1/(eh1+eh2)) in the absence of any regulatory
difference (reference bias due to alignment). The simplest way to get such an estimate would
be to get the median ratio between eh1 and eh2 across the entire library.}

\item{p0}{An average noise rate p(R->A) or p(A->R), i.e., the probability of seeing an allele
due to noise when it is essentially not there (for GTEx v7: LAMP = 0.0003).}
}
\value{
Undajusted 2-sided p-value, testing whether: H0:There does not exist a significantly
abnormal allelic imbalance for this SNP, vs. H1: Allelic ratio significantly deviates from
normal population.
}
\description{
This is a black box function which performs the statistical test on a single SNP.
}
