% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vg_GTEx_v7}
\alias{Vg_GTEx_v7}
\title{Vg_GTEx_v7}
\format{A data frame with 14091 rows and 51 variables:
\describe{
  \item{IDs}{Ensembl ID}
  \item{BRNACC}{}
  \item{ARTCRN}{}
  \item{BRNHPT}{}
  \item{HRTLV}{}
  \item{ADPVSC}{}
  \item{CLNTRN}{}
  \item{PNCREAS}{}
  \item{TESTIS}{}
  \item{LUNG}{}
  \item{ADRNLG}{}
  \item{BREAST}{}
  \item{BLDDER}{}
  \item{SKINS}{}
  \item{SPLEEN}{}
  \item{BRNAMY}{}
  \item{BRNCHB}{}
  \item{BRNCDT}{}
  \item{BRNCTXB}{}
  \item{KDNCTX}{}
  \item{PRSTTE}{}
  \item{BRNHPP}{}
  \item{UTERUS}{}
  \item{CLNSGM}{}
  \item{ESPMCS}{}
  \item{PTTARY}{}
  \item{SKINNS}{}
  \item{ESPGEJ}{}
  \item{ARTTBL}{}
  \item{LCL}{}
  \item{MSCLSK}{}
  \item{SNTTRM}{}
  \item{BRNSNG}{}
  \item{THYROID}{}
  \item{STMACH}{}
  \item{FIBRBLS}{}
  \item{VAGINA}{}
  \item{BRNPTM}{}
  \item{LIVER}{}
  \item{BRNNCC}{}
  \item{BRNSPC}{}
  \item{WHLBLD}{}
  \item{ADPSBQ}{}
  \item{OVARY}{}
  \item{SLVRYG}{}
  \item{HRTAA}{}
  \item{BRNCHA}{}
  \item{ESPMSL}{}
  \item{BRNCTXA}{}
  \item{ARTAORT}{}
  \item{NERVET}{}
}}
\source{
\url{https://gtexportal.org/home/}
}
\usage{
Vg_GTEx_v7
}
\description{
A dataset containing the estimated variance in gene expression obtained
from the GTEx project.
}
\details{
The Genotype-Tissue Expression (GTEx) Project was supported by the Common
Fund of the Office of the Director of the National Institutes of Health,
and by NCI, NHGRI, NHLBI, NIDA, NIMH, and NINDS. The data used for the
analyses described in this manuscript were obtained from the GTEx Portal
on 02/17/19.
}
\keyword{datasets}
