% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVAdot.R, R/ASEoutlier.R
\name{ASEoutlier}
\alias{ASEoutlier}
\alias{ASEoutlier}
\title{ANEVA Dosage Outlier Test}
\usage{
ASEoutlier(filepath_ASE, output_columns = c("refCount", "altCount"),
  eh1 = "refCount", eh2 = "altCount", ENS_id = "name", filepath_Vg, Vg,
  FDR = 0.05, plot = TRUE)

ASEoutlier(filepath_ASE, output_columns = c("refCount", "altCount"),
  eh1 = "refCount", eh2 = "altCount", ENS_id = "name", filepath_Vg, Vg,
  FDR = 0.05, plot = TRUE)
}
\arguments{
\item{filepath_ASE}{A string with quotation marks around it indicating ASE count data filepath}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{ENS_id}{A string with the column name of the ensemble id associated with the count data}

\item{filepath_Vg}{A string with quotation marks around it indicating filepath for Variance estimates}

\item{Vg}{a vector with variance estimates for dosage distributions associated with count data.}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}

\item{plot}{A logical T/F indicating whether plots should be generated for the test data}

\item{filepath}{A string with quotation marks around it indicating ASE count data filepath}

\item{output_columnns}{A vector containing strings of the exact column names to be duplicated
from input file in the output}

\item{output_columnns}{A vector containing strings of the exact column names to be duplicated
from input file in the output}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{Vg}{a string indicating the column name where the variances can be found in the Variance file}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}

\item{plot}{A logical T/F indicating whether plots should be generated for the test data}
}
\value{
A table containing the user-specified output columns, as well as unadjusted and
Benjamini-Hoschberg adjusted p-values for detection of potential dosage outlier.

A table containing the user-specified output columns, as well as unadjusted and
Benjamini-Hoschberg adjusted p-values for h1M, h2M and total change of expression.
}
\description{
This test is designed to detect if ASE data reveals sufficient imbalance to induce
an outlier in total gene expression.

This test is designed to detect if ASE shows sufficient imbalance to induce
an outlier in total gene expression. We first test under the assumption that
the mutation lies on haplotype 1, then test again assuming that the
mutation lies on haplotype 2. These results are averaged, resulting in an
effective two-sided test.
}
\examples{
(tbd)
(tbd)
}
