% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEoutlier.R
\name{ASEoutlier}
\alias{ASEoutlier}
\title{Bayesian Change of Expression test on ASE count data.}
\usage{
ASEoutlier(filepath, output_columns = c("refCount", "altCount"), r,
  eh1 = "refCount", eh2 = "altCount", Vg, FDR = 0.05, plot = TRUE)
}
\arguments{
\item{filepath}{A string with quotation marks around it indicating ASE data filepath}

\item{r}{The ?true reference allele ratio?}

\item{eh1}{A string with the column name of the reference count data}

\item{eh2}{A string with the column name of the alternative count data}

\item{Vg}{A numeric value for the estimated dosage variance in normal population}

\item{FDR}{A numeric value between 0 and 1 indicating the desired false discovery rate}

\item{plot}{A logical T/F indicating whether plots should be generated for the test data}

\item{output_columnns}{A vector containing strings of the exact column names to be duplicated
from input file in the output}
}
\value{
A table containing the user-specified output columns, as well as unadjusted and
Benjamini-Hoschberg adjusted p-values for Loss, Gain, and change of expression
}
\description{
Bayesian Change of Expression test on ASE count data.
}
\examples{
(tbd)
}
