% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVAdot.R, R/ASEoutlier.R
\name{integrand.h1M}
\alias{integrand.h1M}
\alias{integrand.h1M}
\title{We implement a helper function to generate the desired integrand
for the h1M test, that is, assuming that haplotype1 harbors the mutation.}
\usage{
integrand.h1M(r, eh1, eh2, Vg)

integrand.h1M(r, eh1, eh2, Vg)
}
\arguments{
\item{r}{Variable representing the true allelic ratio of haplotype 1}

\item{eh1}{An integer value for count of expression of haplotype1}

\item{eh2}{An integer value for count of expression of haplotype2}

\item{Vg}{A numeric value for the estimated dosage variance in normal population
(can be gotten from GTEx)}

\item{r}{Variable representing the true allelic ratio of haplotype 1}

\item{eh1}{An integer value for count of expression of haplotype1}

\item{eh2}{An integer value for count of expression of haplotype2}

\item{Vg}{A numeric value for the estimated dosage variance in normal population
(can be gotten from GTEx)}
}
\value{
The integrand for which we wish to integrate over to get
desired p-values

The integrand for which we wish to integrate over to get
desired p-values
}
\description{
We implement a helper function to generate the desired integrand
for the h1M test, that is, assuming that haplotype1 harbors the mutation.

We implement a helper function to generate the desired integrand
for the h1M test, that is, assuming that haplotype1 harbors the mutation.
}
\examples{
(tbd)
(tbd)
}
