% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANEVAdot.R
\name{integrand}
\alias{integrand}
\title{We implement a helper function to generate the desired integrand
for the test.}
\usage{
integrand(dE, eh1, eh2, Eg_std)
}
\arguments{
\item{dE}{Variable representing the true change in gene dosage
(variable of integration)}

\item{eh1}{An integer value for count of expression of haplotype1}

\item{eh2}{An integer value for count of expression of haplotype2}

\item{Eg_std}{A numeric value for the estimated dosage standard deviation
in normal population (can be derived from GTEx)}
}
\value{
The integrand over which we wish to integrate to get
desired p-values
}
\description{
We implement a helper function to generate the desired integrand
for the test.
}
